+++ source/SXGUIWinApi/sxgui_button.h
@@ -11,7 +11,7 @@ class SXGUIButton : public SXGUITextual, public virtual ISXGUIButton
 public:
 	SXGUIButton();
 	SXGUIButton(const char* caption,WORD x,WORD y,WORD width,WORD heigth,DWORD exstyle,DWORD style,HWND parent,WNDPROC handler,DWORD id);
-	SXGUIButton(const char* caption,WORD x,WORD y,WORD width,WORD heigth,WORD image,HWND parent,WNDPROC handler,DWORD id);
+	SXGUIButton(const char* caption, WORD x, WORD y, WORD width, WORD heigth, int image, HWND parent, WNDPROC handler, DWORD id);
 
 	bool SetIconInFile(const char* path);
 	bool SetBmpInFile(const char* path);
